Orderly Location for Sharepoint

A driver to put orderly/orderly2 archives on Sharepoint. Individual methods are not documented here as this just needs to satisfy the interface required by orderly/orderly2 and is not for direct user use. See orderly2::orderly_location_add for details.

Methods

Public methods

orderly_location_sharepoint$new()

orderly_location_sharepoint$list()

orderly_location_sharepoint$metadata()

orderly_location_sharepoint$fetch_file()

orderly_location_sharepoint$list_unknown_packets()

orderly_location_sharepoint$list_unknown_files()

orderly_location_sharepoint$push_file()

orderly_location_sharepoint$push_metadata()

orderly_location_sharepoint$clone()

Method new()

Usage

orderly_location_sharepoint$new(url, site, path)

Arguments

url

The base url of your Office365/Sharepoint site, such as https://myorg.sharepoint.com (the https:// prefix is required)

site

Your site name on Sharepoint

path

The path within your site name where documents will be stored (you may need Shared Documents even if Sharepoint makes it look like Documents

Method list()

Usage

orderly_location_sharepoint$list()

Method metadata()

Usage

orderly_location_sharepoint$metadata(packet_ids)

Method fetch_file()

Usage

orderly_location_sharepoint$fetch_file(hash, dest)

Method list_unknown_packets()

Usage

orderly_location_sharepoint$list_unknown_packets(ids)

Method list_unknown_files()

Usage

orderly_location_sharepoint$list_unknown_files(hashes)

Method push_file()

Usage

orderly_location_sharepoint$push_file(src, hash)

Method push_metadata()

Usage

orderly_location_sharepoint$push_metadata(packet_id, root)

Method clone()

The objects of this class are cloneable with this method.

Usage

orderly_location_sharepoint$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.